Refactor branch treeview to separate file
